file name changes, moving map into game page
